Added the environment variable and URL encoding placeholders
